Extrator/requirements.txt
bcrypt==4.0.1
cffi==1.15.1
cryptography==41.0.1
greenlet==2.0.2
paramiko==3.2.0
psycopg2==2.9.6
pycparser==2.21
PyNaCl==1.5.0
python-dotenv==1.0.0
SQLAlchemy==2.0.16
sshtunnel==0.4.0
typing_extensions==4.6.3
